requirements.txt
altair-ally==0.1.1
vega-datasets==0.9.0
vegafusion==1.6.9
deepchecks==0.18.1
pandera==0.20.4
tabulate==0.9.0
click==8.1.7
vl_convert_python==1.7.0
pytest==8.3.4
